Move getGSCMembers to own file
