+++ apps/council-ui/src/ui/vaults/gscVault/GSCVaultDetails.tsx
@@ -12,7 +12,10 @@ import { GSCVaultStatsBar } from "src/ui/vaults/gscVault/GSCVaultStatsBar/GSCVau
 import { GSCVaultStatsBarSkeleton } from "src/ui/vaults/gscVault/GSCVaultStatsBar/GSCVaultStatsBarSkeleton";
 
 import { VaultHeader, VaultHeaderSkeleton } from "src/ui/vaults/VaultHeader";
-import { getGSCMembers, GSCMemberInfo } from "src/vaults/gscVault";
+import {
+  getGSCMembers,
+  GSCMemberInfo,
+} from "src/vaults/gscVault/getGSCMembers";
 import { getGSCStatus } from "src/vaults/gscVault/getGSCStatus";
 import { GSCStatus } from "src/vaults/gscVault/types";
 import { useAccount } from "wagmi";
